refactor: extract Terminal component from web channel
The web channel's app.tsx contained both the Terminal UI logic and the
page entry point in a single file. Extract the reusable Terminal
component (with its hooks, helpers, and CSS) into terminal.tsx and
terminal.css so that the gateway's mux UI can share the same component
without duplicating code.
